// src/types/invidious.ts
export interface ThumbnailObject {
  quality: string;
  url: string;
  width: number;
  height: number;
}

export interface ImageObject {
  url: string;
  width: number;
  height: number;
}

export interface VideoObject {
  type: 'video';
  title: string;
  videoId: string;
  author: string;
  authorId: string;
  authorUrl: string;
  authorVerified?: boolean;
  videoThumbnails: ThumbnailObject[];
  description?: string;
  descriptionHtml?: string;
  viewCount?: number;
  viewCountText?: string;
  lengthSeconds?: number;
  published?: number;
  publishedText?: string;
  premiereTimestamp?: number;
  liveNow?: boolean;
  premium?: boolean;
  isUpcoming?: boolean;
  isNew?: boolean;
  is4k?: boolean;
  is8k?: boolean;
  hasCaptions?: boolean;
}

export interface ChannelObject {
  type: 'channel';
  author: string;
  authorId: string;
  authorUrl: string;
  authorVerified?: boolean;
  authorThumbnails: ThumbnailObject[];
  autoGenerated?: boolean;
  subCount?: number;
  videoCount?: number;
  description?: string;
  descriptionHtml?: string;
}

export interface PlaylistObject {
  type: 'playlist';
  title: string;
  playlistId: string;
  playlistThumbnail?: string;
  author: string;
  authorId: string;
  authorUrl: string;
  authorVerified?: boolean;
  videoCount?: number;
  videos: {
    title: string;
    videoId: string;
    lengthSeconds: number;
    videoThumbnails: ThumbnailObject[];
  }[];
}
// ...
export interface ChannelDetails {
  author: string;
  authorId: string;
  authorUrl: string;
  authorVerified: boolean;
  authorBanners: ImageObject[];
  authorThumbnails: ThumbnailObject[];
  subCount: number;
  totalViews: number;
  joined: number;
  autoGenerated: boolean;
  isFamilyFriendly: boolean;
  description: string;
  descriptionHtml: string;
  allowedRegions: string[];
  tabs: string[];
  pronouns?: string;
  latestVideos: VideoObject[];
  relatedChannels: ChannelObject[];
}
// ...
export interface CommunityPost {
  attachment?: unknown;
  author: string;
  authorIsChannelOwner: boolean;
  authorId: string;
  authorThumbnails: ImageObject[];
  authorUrl: string;
  commentId: string;
  content: string;
  contentHtml: string;
  isEdited: boolean;
  likeCount: number;
  published: number;
  publishedText: string;
  replyCount: number;
}